added folders and readmes for most folders to clarify usage. moved css to assets
