+++ Frontend/src/App.jsx
@@ -1,4 +1,4 @@
-import './App.css'
+import './assets/App.css'
 
 function App() {
   return (
